refactor Ray class into its own file
